refactor: move key bundle parsing tests to pkg/signatures
Tests for ParseKeyBundle, PEM canonicalization, and
BundleToSignatureIntegration belong alongside the code they test in
pkg/signatures, not in the central datastore package.

Handler/watcher tests that use mock stores remain in central.

Partially generated by AI.
